Moved health script
Player health stuff is now in a separate script

If you're merging you can ignore all changes to CharacterControl
