+++ Assets/Scenes/Enemy fireball.unity
@@ -696,6 +696,7 @@ GameObject:
   - component: {fileID: 1051865577275999239}
   - component: {fileID: 4935163188144231719}
   - component: {fileID: 7992123443317296978}
+  - component: {fileID: 1513896016}
   - component: {fileID: 1513896017}
   m_Layer: 9
   m_Name: default_sprite
@@ -768,6 +769,19 @@ Transform:
   m_Father: {fileID: 2006574258452901336}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1513896016
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1513896013}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ee8e89ad94d0e7748a0db3b0a3306df1, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  healthText: {fileID: 393016118}
 --- !u!95 &1513896017
 Animator:
   serializedVersion: 3
@@ -21520,7 +21534,7 @@ Tilemap:
   - first: {x: 46, y: 3, z: 0}
     second:
       m_TileIndex: 17
-      m_TileSpriteIndex: 17
+      m_TileSpriteIndex: 18
       m_TileMatrixIndex: 1
       m_TileColorIndex: 0
       m_ObjectToInstantiate: {fileID: 0}
@@ -23810,10 +23824,10 @@ Tilemap:
     m_Data: {fileID: 0}
   - m_RefCount: 0
     m_Data: {fileID: 0}
-  - m_RefCount: 1
-    m_Data: {fileID: 21300016, guid: d4024c3b29a5c3545b31c279f44e19f5, type: 3}
   - m_RefCount: 0
     m_Data: {fileID: 0}
+  - m_RefCount: 1
+    m_Data: {fileID: 21300016, guid: d4024c3b29a5c3545b31c279f44e19f5, type: 3}
   - m_RefCount: 3
     m_Data: {fileID: 21300098, guid: d4024c3b29a5c3545b31c279f44e19f5, type: 3}
   - m_RefCount: 93
@@ -23984,7 +23998,6 @@ MonoBehaviour:
     jumpDelay: 0.1
     gravity: 9.81
   footPos: {fileID: 8457548506199831824}
-  healthText: {fileID: 393016118}
 --- !u!4 &8457548506199831824
 Transform:
   m_ObjectHideFlags: 0
