+++ Assets/Scripts/CharacterHealth.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ee8e89ad94d0e7748a0db3b0a3306df1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
